requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
Pillow==10.1.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
openai==1.3.5
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
geopy==2.4.0
django-extensions==3.2.3
celery==5.3.4
redis==5.0.1
stripe==7.8.0
jellyfish==1.0.3
django-unfold==0.23.0
